% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_lakes}
\alias{major_lakes}
\title{Lake Malawi Boundary}
\format{
An sf object with 1 feature and Natural Earth lake attributes:
\describe{
\item{name}{Lake name}
\item{name_alt}{Alternative lake name}
\item{geometry}{MULTIPOLYGON geometry for the lake boundary}
}
}
\source{
Natural Earth lake polygons
}
\usage{
major_lakes
}
\description{
Spatial data for Lake Malawi, also known as Lake Nyasa.
}
\keyword{datasets}
