% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_level_0}
\alias{mw_level_0}
\title{Malawi Administrative Boundaries - Level 0 (Country)}
\format{
An sf object with 1 feature and 5 fields:
\describe{
\item{ADM0_EN}{Country name (Malawi)}
\item{ADM0_PCODE}{Country ISO code (MW)}
\item{area_sqkm}{Area in square kilometres}
\item{center_lat}{Centroid latitude}
\item{center_lon}{Centroid longitude}
\item{geometry}{MULTIPOLYGON geometry for the national boundary}
}
}
\source{
Malawi Spatial Data Platform (MASDAP) / OCHA HDX, version 2 (2023)
}
\usage{
mw_level_0
}
\description{
National boundary of Malawi at administrative level 0.
}
\keyword{datasets}
