% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_level_1}
\alias{mw_level_1}
\title{Malawi Administrative Boundaries - Level 1 (Regions)}
\format{
An sf object with 3 features and 7 fields:
\describe{
\item{ADM1_EN}{Region name (Northern, Central, Southern)}
\item{ADM1_PCODE}{Region code (MW1, MW2, MW3)}
\item{ADM0_EN}{Country name}
\item{ADM0_PCODE}{Country code}
\item{area_sqkm}{Area in square kilometres}
\item{center_lat}{Centroid latitude}
\item{center_lon}{Centroid longitude}
\item{geometry}{MULTIPOLYGON geometry for regional boundaries}
}
}
\source{
Malawi Spatial Data Platform (MASDAP) / OCHA HDX, version 2 (2023)
}
\usage{
mw_level_1
}
\description{
The three administrative regions of Malawi: Northern, Central, and Southern.
}
\keyword{datasets}
