% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mwmapdata_check}
\alias{mwmapdata_check}
\title{Check if mwmapdata spatial data is valid}
\usage{
mwmapdata_check(x = mw_level_2)
}
\arguments{
\item{x}{An sf object to check. Defaults to \code{mw_level_2}.}
}
\value{
Logical. \code{TRUE} if \code{x} is a valid sf object.
}
\description{
A utility to confirm that the spatial objects in \pkg{mwmapdata} are
well-formed sf objects.
}
