% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major_lakes}
\alias{major_lakes}
\title{Major Lakes of Malawi}
\format{
An sf object with 3 features and 3 fields:
\describe{
\item{name}{Lake name}
\item{area_km2}{Approximate surface area in square kilometres}
\item{max_depth_m}{Maximum depth in metres (where available)}
\item{geometry}{MULTIPOLYGON geometry for lake boundaries}
}
}
\source{
Malawi Spatial Data Platform (MASDAP) / Department of Surveys
}
\usage{
major_lakes
}
\description{
Spatial data for the major lakes in Malawi:
Lake Malawi (Lake Nyasa), Lake Malombe, and Lake Chilwa.
}
\keyword{datasets}
