% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malawi_data}
\alias{malawi_data}
\title{Legacy Malawi States Dataset}
\format{
An sf object (deprecated)
}
\usage{
malawi_data
}
\description{
\strong{Deprecated.} Legacy dataset containing Malawi states.
Please use \code{mw_level_2} instead.
}
\keyword{internal}
